Rename Class and move package
